requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
Pillow>=10.4.0
pillow-heif>=0.16.0
pypdf>=4.0.1
google-genai>=0.4.0
httpx>=0.27.0
stripe>=11.1.0
